requirement.txt
asgiref==3.6.0
attrs==22.2.0
click==8.1.3
colorama==0.4.6
databases==0.7.0
Django==4.2
exceptiongroup==1.1.1
Flask==2.2.3
flask-expects-json==1.7.0
Flask-Inputs==0.3.0
future==0.18.3
greenlet==2.0.2
image==1.5.33
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonify==0.5
jsonschema==4.17.3
MarkupSafe==2.1.2
marshmallow==3.19.0
orm==0.3.1
packaging==23.0
Pillow==9.5.0
pluggy==1.0.0
pyrsistent==0.19.3
pytest==7.3.1
six==1.16.0
SQLAlchemy==1.4.47
sqlparse==0.4.3
tomli==2.0.1
typesystem==0.3.1
tzdata==2023.3
Werkzeug==2.2.3
WTForms==3.0.1